requirements.txt
absl-py==2.1.0
alembic==1.13.1
annotated-types==0.6.0
anyio==4.2.0
astunparse==1.6.3
bcrypt==4.1.2
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cryptography==42.0.2
cycler==0.12.1
dm-tree==0.1.8
ecdsa==0.18.0
fastapi==0.108.0
flatbuffers==23.5.26
fonttools==4.47.2
gast==0.5.4
google-auth==2.26.2
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.60.0
gunicorn==21.2.0
h11==0.14.0
h5py==3.10.0
httptools==0.6.1
idna==3.6
joblib==1.3.2
keras==2.15.0
kiwisolver==1.4.5
libclang==16.0.6
Mako==1.3.0
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
mdurl==0.1.2
ml-dtypes==0.2.0
namex==0.0.7
numpy==1.26.3
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.4
passlib==1.7.4
pillow==10.2.0
protobuf==4.23.4
psycopg2==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.17.2
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rsa==4.9
scikit-learn==1.4.0
scipy==1.11.4
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.32.0.post1
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
threadpoolctl==3.2.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
uvicorn==0.25.0
waitress==2.1.2
watchfiles==0.21.0
websockets==12.0
Werkzeug==3.0.1
wrapt==1.14.1
